requirements.txt
requests
pandas
python-dotenv
black
nltk
openpyxl
ruff